moved init registry fixture to conftest
