+++ services/director/tests/conftest.py
@@ -2,6 +2,11 @@ import pytest
 import logging
 from pathlib import Path
 import sys
+
+from simcore_service_director import (
+    config,
+    registry_proxy
+    )
 # pylint:disable=unused-argument
 
 pytest_plugins = ["fixtures.docker_registry"]
@@ -14,3 +19,10 @@ CURRENT_DIR = Path(sys.argv[0] if __name__ == "__main__" else __file__).parent.a
 def docker_compose_file(pytestconfig):
     my_path = CURRENT_DIR / "docker-compose.yml"
     return my_path
+
+@pytest.fixture
+def configure_registry_access(docker_registry):
+    config.REGISTRY_URL = docker_registry
+    config.REGISTRY_SSL = False
+    config.CONVERT_OLD_API = False
+    registry_proxy.setup_registry_connection()
