+++ services/director/tests/test_registry_proxy.py
@@ -7,12 +7,6 @@ from simcore_service_director import (
     )
 from simcore_service_director.exceptions import DirectorException
 
-@pytest.fixture
-def proxy_setup(docker_registry):
-    config.REGISTRY_URL = docker_registry
-    config.REGISTRY_SSL = False
-    registry_proxy.setup_registry_connection()
-
 def test_setup_registry_connection():
     config.REGISTRY_AUTH = False
     try:
@@ -29,23 +23,23 @@ def test_setup_registry_connection():
     config.REGISTRY_PW = "TheUgly"
     registry_proxy.setup_registry_connection()
 
-def test_list_no_services_available(docker_registry, proxy_setup):    
+def test_list_no_services_available(docker_registry, configure_registry_access):    
     computational_services = registry_proxy.list_computational_services()
     assert (not computational_services) # it's empty
     interactive_services = registry_proxy.list_interactive_services()
     assert (not interactive_services)
 
-def test_list_computational_services(docker_registry, push_services, proxy_setup):
+def test_list_computational_services(docker_registry, push_services, configure_registry_access):
     push_services(6, 3)    
     computational_services = registry_proxy.list_computational_services()
     assert len(computational_services) == 6
 
-def test_list_interactive_services(docker_registry, push_services, proxy_setup):
+def test_list_interactive_services(docker_registry, push_services, configure_registry_access):
     push_services(5, 4)    
     interactive_services = registry_proxy.list_interactive_services()
     assert len(interactive_services) == 4
 
-def test_retrieve_list_of_images_in_repo(docker_registry, push_services, proxy_setup):
+def test_retrieve_list_of_images_in_repo(docker_registry, push_services, configure_registry_access):
     images = push_services(5, 3)
     image_number = {}
     for image in images:
@@ -64,7 +58,7 @@ def test_list_interactive_service_dependencies():
     # need to setup a fake registry to test this
     pass
 
-def test_retrieve_labels_of_image(docker_registry, push_services, proxy_setup):
+def test_retrieve_labels_of_image(docker_registry, push_services, configure_registry_access):
     images = push_services(1, 1)    
     for image in images:
         service_description = image["service_description"]
@@ -115,7 +109,7 @@ def test_get_service_last_namess():
     repo = "services/dynamic/modeler"
     assert registry_proxy.get_service_last_names(repo) == "invalid service"
 
-def test_get_service_details(push_services, proxy_setup):
+def test_get_service_details(push_services, configure_registry_access):
     images = push_services(1, 1)    
     for image in images:
         service_description = image["service_description"]
